requirements.txt
Pillow==7.1.1
numpy==1.17.4
mtcnn==0.1.0
matplotlib==3.1.2
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
sklearn==0.0
scikit_image==0.16.2
librosa==0.7.2
soundfile==0.10.2
sndfile==0.2.0
numba==0.48.0
pysndfx==0.3.6
tensorflow==2.2.0rc3